# Performance regression eager load