Small PR, just removes the eager loading from framework to speed up boot times, there was a regression when we moved to Rails 7 that enabled eager loading with zeitwerk so it's just reverting that

This should have no effect on Pro since it has eager loading set to [true](https://github.com/rapid7/pro/blob/a261dc5094cd394e33332d45e0a8f1264c717781/ui/config/environments/production.rb#L53) which the docs say will also eager load all gems with zeitwerk https://guides.rubyonrails.org/autoloading_and_reloading_constants.html#eager-loading

I also did boot up pro with eager loading turned off on framework and I didn't run into any issues

I'm also namespacing some `print_*` methods in `lib/msf/core/web_services/http_db_manager_service.rb` when eager loading these were available globally and accidentally being called, now that it's no longer being eagerly loaded I namespaced it's `print_*` methods and ran through all the modules to ensure they still had access to the various `print_*` methods and I didn't find any issues there either

# Verification Steps
- [ ] On master `time ./msfconsole -qx 'exit'`
- [ ] On this PR `time ./msfconsole -qx 'exit'`
- [ ] There should be a noticeable time save
- [ ] CI tests pass

Using multitime for multiple runs I got these results
<img width="568" alt="image" src="https://user-images.githubusercontent.com/19910435/225611917-9043a5e0-554c-47e3-a710-3ea032ca0481.png">
On master:
```
$ multitime -n 5 ./msfconsole -qx 'exit'                                                    
===> multitime results
1: ./msfconsole -qx exit
            Mean        Std.Dev.    Min         Median      Max
real        14.743      3.060       11.354      14.879      19.399
user        9.216       0.897       8.406       9.075       10.892
sys         3.346       0.271       3.046       3.344       3.805
```

On this PR:
```
$ multitime -n 5 ./msfconsole -qx 'exit'                                                    
===> multitime results
1: ./msfconsole -qx exit
            Mean        Std.Dev.    Min         Median      Max
real        9.067       0.438       8.731       8.892       9.934
user        6.544       0.055       6.448       6.553       6.618
sys         2.375       0.020       2.337       2.379       2.396
```