@dwelch-r7 2023-03-16
Note to self add in a test to make sure that all autoloaded files can still be loaded